refactor #4
