+++ packages/fuels-core/src/code_gen/custom_types_gen/mod.rs
@@ -1,424 +1,10 @@
-use crate::code_gen::resolved_type::{resolve_type, ResolvedType};
-use crate::utils::ident;
-use crate::{try_from_bytes, ParamType, Parameterize, Token, Tokenizable};
-use anyhow::{anyhow, bail};
-use fuels_types::errors::Error;
-use fuels_types::param_types::EnumVariants;
-use fuels_types::utils::has_array_format;
-use fuels_types::{CustomType, TypeApplication, TypeDeclaration};
-use inflector::Inflector;
-use itertools::Itertools;
-use lazy_static::lazy_static;
-use proc_macro2::{Ident, LexError, TokenStream};
-use quote::{quote, ToTokens};
-use regex::{Captures, Regex};
-use std::collections::{HashMap, HashSet};
-use std::iter::Map;
-use std::slice::Iter;
-use std::str::FromStr;
-use syn::parse_macro_input;
+mod common;
+mod enum_gen;
+mod struct_gen;
 
-struct Component {
-    pub field_name: Ident,
-    pub field_type: ResolvedType,
-}
-
-impl Component {
-    pub fn new(
-        component: &TypeApplication,
-        types: &HashMap<usize, TypeDeclaration>,
-        snake_case: bool,
-    ) -> anyhow::Result<Component> {
-        let field_name = if snake_case {
-            component.name.to_snake_case()
-        } else {
-            component.name.to_owned()
-        };
-
-        Ok(Component {
-            field_name: ident(&field_name),
-            field_type: resolve_type(component, types)?,
-        })
-    }
-}
-
-pub fn expand_custom_struct(
-    prop: &TypeDeclaration,
-    types: &HashMap<usize, TypeDeclaration>,
-) -> Result<TokenStream, Error> {
-    let struct_ident = extract_custom_type_name_from_abi_property(prop)?;
-
-    let field_entries = extract_components(&prop, types, true)?;
-    let generic_parameters = extract_generic_parameters(&field_entries)?;
-
-    let struct_decl = struct_decl(&struct_ident, &field_entries, &generic_parameters);
-
-    let parameterized_impl =
-        struct_parameterized_impl(&field_entries, &struct_ident, &generic_parameters);
-
-    let tokenizable_impl =
-        struct_tokenizable_impl(&struct_ident, &field_entries, &generic_parameters);
-
-    let try_from = impl_try_from(&struct_ident, &generic_parameters);
-
-    Ok(quote! {
-        #struct_decl
-
-        #parameterized_impl
-
-        #tokenizable_impl
-
-        #try_from
-    })
-}
-
-pub fn expand_custom_enum(
-    prop: &TypeDeclaration,
-    types: &HashMap<usize, TypeDeclaration>,
-) -> Result<TokenStream, Error> {
-    let enum_ident = &extract_custom_type_name_from_abi_property(prop)?;
-
-    let field_entries = extract_components(&prop, types, false)?;
-    let generics = extract_generic_parameters(&field_entries)?;
-
-    let enum_def = enum_decl(&enum_ident, &field_entries, &generics);
-    let parameterize_impl = enum_parameterize_impl(&enum_ident, &field_entries, &generics);
-    let tokenize_impl = enum_tokenizable_impl(&enum_ident, &field_entries, &generics);
-    let try_from = impl_try_from(&enum_ident, &generics);
-
-    Ok(quote! {
-        #enum_def
-
-        #parameterize_impl
-
-        #tokenize_impl
-
-        #try_from
-    })
-}
-
-fn impl_try_from(ident: &Ident, generics: &[TokenStream]) -> TokenStream {
-    quote! {
-        impl<#(#generics: Tokenizable + Parameterize,)*> TryFrom<&[u8]> for #ident<#(#generics,)*> {
-            type Error = SDKError;
-
-            fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-                try_from_bytes(bytes)
-            }
-        }
-        impl<#(#generics: Tokenizable + Parameterize,)*> TryFrom<&Vec<u8>> for #ident<#(#generics,)*> {
-            type Error = SDKError;
-
-            fn try_from(bytes: &Vec<u8>) -> Result<Self, Self::Error> {
-                try_from_bytes(&bytes)
-            }
-        }
-
-        impl<#(#generics: Tokenizable + Parameterize,)*> TryFrom<Vec<u8>> for #ident<#(#generics,)*> {
-            type Error = SDKError;
-
-            fn try_from(bytes: Vec<u8>) -> Result<Self, Self::Error> {
-                try_from_bytes(&bytes)
-            }
-        }
-    }
-}
-
-fn enum_decl(
-    enum_ident: &Ident,
-    field_entries: &[Component],
-    generics: &[TokenStream],
-) -> TokenStream {
-    let enum_variants = field_entries.iter().map(
-        |Component {
-             field_name,
-             field_type,
-         }| {
-            let field_type = if let ParamType::Unit = field_type.param_type {
-                quote! {}
-            } else {
-                field_type.into()
-            };
-
-            quote! {
-                #field_name(#field_type)
-            }
-        },
-    );
-
-    quote! {
-        #[derive(Clone, Debug, Eq, PartialEq)]
-        pub enum #enum_ident <#(#generics: Tokenizable + Parameterize,)*> {
-            #(#enum_variants,)*
-        }
-    }
-}
-
-fn enum_tokenizable_impl(
-    enum_ident: &Ident,
-    field_entries: &[Component],
-    generics: &[TokenStream],
-) -> TokenStream {
-    let enum_name = enum_ident.to_string();
-
-    let match_discriminant_from_token = field_entries.iter().enumerate().map(
-        |(
-            discriminant,
-            Component {
-                field_name,
-                field_type,
-            },
-        )| {
-            let value = if let ParamType::Unit = field_type.param_type {
-                quote! {}
-            } else {
-                let field_type: TokenStream = field_type.into();
-                quote! { <#field_type>::from_token(variant_token)? }
-            };
-
-            let u8_discriminant = discriminant as u8;
-            quote! { #u8_discriminant => Ok(Self::#field_name(#value))}
-        },
-    );
-
-    let match_discriminant_into_token = field_entries.iter().enumerate().map(
-        |(
-            discriminant,
-            Component {
-                field_name,
-                field_type,
-            },
-        )| {
-            let u8_discriminant = discriminant as u8;
-            if let ParamType::Unit = field_type.param_type {
-                quote! { Self::#field_name() => (#u8_discriminant, ().into_token())}
-            } else {
-                quote! { Self::#field_name(inner) => (#u8_discriminant, inner.into_token())}
-            }
-        },
-    );
-
-    quote! {
-            impl<#(#generics: Tokenizable + Parameterize,)*> Tokenizable for #enum_ident <#(#generics,)*> {
-                fn from_token(token: Token) -> Result<Self, SDKError>
-                where
-                    Self: Sized,
-                {
-                    let gen_err = |msg| {
-                        SDKError::InvalidData(format!(
-                            "Error while instantiating {} from token! {}", #enum_name, msg
-                        ))
-                    };
-                    match token {
-                        Token::Enum(selector) => {
-                            let (discriminant, variant_token, _) = *selector;
-                            match discriminant {
-                                #(#match_discriminant_from_token,)*
-                                _ => Err(gen_err(format!(
-                                    "Discriminant {} doesn't point to any of the enums variants.", discriminant
-                                ))),
-                            }
-                        }
-                        _ => Err(gen_err(format!(
-                            "Given token ({}) is not of the type Token::Enum!", token
-                        ))),
-                    }
-                }
-
-                fn into_token(self) -> Token {
-                    let (discriminant, token) = match self {
-                        #(#match_discriminant_into_token,)*
-                    };
-
-                    let variants = match Self::param_type() {
-                        ParamType::Enum(variants) => variants,
-                        other => panic!("Calling {}::param_type() must return a ParamType::Enum but instead it returned: {}", #enum_name, other)
-                    };
-
-                    Token::Enum(Box::new((discriminant, token, variants)))
-                }
-            }
-    }
-}
-
-fn enum_parameterize_impl(
-    enum_ident: &Ident,
-    field_entries: &[Component],
-    generics: &[TokenStream],
-) -> TokenStream {
-    let param_type_calls = param_type_calls(&field_entries);
-    let enum_name = enum_ident.to_string();
-    quote! {
-        impl<#(#generics: Parameterize + Tokenizable,)*> Parameterize for #enum_ident <#(#generics,)*> {
-            fn param_type() -> ParamType {
-                let mut param_types = vec![];
-                #(param_types.push(#param_type_calls);)*
-
-                let variants = EnumVariants::new(param_types).unwrap_or_else(|_| panic!("{} has no variants which isn't allowed!", #enum_name));
-                ParamType::Enum(variants)
-            }
-        }
-    }
-}
-
-fn struct_decl(
-    struct_ident: &Ident,
-    field_entries: &Vec<Component>,
-    generic_parameters: &Vec<TokenStream>,
-) -> TokenStream {
-    let fields = field_entries.iter().map(
-        |(Component {
-             field_name,
-             field_type,
-         })| {
-            let field_type: TokenStream = field_type.into();
-            quote! { pub #field_name: #field_type }
-        },
-    );
-
-    quote! {
-        #[derive(Clone, Debug, Eq, PartialEq)]
-        pub struct #struct_ident <#(#generic_parameters: Tokenizable + Parameterize, )*> {
-            #(#fields),*
-        }
-    }
-}
-
-fn struct_tokenizable_impl(
-    struct_ident: &Ident,
-    field_entries: &[Component],
-    generic_parameters: &Vec<TokenStream>,
-) -> TokenStream {
-    let struct_name_str = struct_ident.to_string();
-    let from_token_calls = field_entries
-        .iter()
-        .map(
-            |(Component {
-                 field_name,
-                 field_type,
-             })| {
-                let resolved: TokenStream = field_type.into();
-                quote! {
-                    #field_name: <#resolved>::from_token(next_token()?)?
-                }
-            },
-        )
-        .collect::<Vec<_>>();
-
-    let into_token_calls = field_entries
-        .iter()
-        .map(|Component { field_name, .. }| {
-            quote! {self.#field_name.into_token()}
-        })
-        .collect::<Vec<_>>();
-
-    quote! {
-        impl <#(#generic_parameters: Tokenizable + Parameterize, )*> Tokenizable for #struct_ident <#(#generic_parameters, )*> {
-            fn into_token(self) -> Token {
-                let mut tokens = Vec::new();
-                #( tokens.push(#into_token_calls); )*
-                Token::Struct(tokens)
-            }
-
-            fn from_token(token: Token)  -> Result<Self, SDKError> {
-                match token {
-                    Token::Struct(tokens) => {
-                        let mut tokens_iter = tokens.into_iter();
-                        let mut next_token = move || { tokens_iter
-                            .next()
-                            .ok_or_else(|| { SDKError::InstantiationError(format!("Ran out of tokens before '{}' has finished construction!", #struct_name_str)) })
-                        };
-                        Ok(Self { #( #from_token_calls, )* })
-                    },
-                    other => Err(SDKError::InstantiationError(format!("Error while constructing '{}'. Expected token of type Token::Struct, got {:?}", #struct_name_str, other))),
-                }
-            }
-        }
-    }
-}
-
-fn struct_parameterized_impl(
-    field_entries: &[Component],
-    struct_ident: &Ident,
-    generic_parameters: &[TokenStream],
-) -> TokenStream {
-    let param_type_calls = param_type_calls(&field_entries);
-    quote! {
-        impl <#(#generic_parameters: Parameterize + Tokenizable,)*> Parameterize for #struct_ident <#(#generic_parameters,)*> {
-            fn param_type() -> ParamType {
-                let mut types = Vec::new();
-                #( types.push(#param_type_calls); )*
-                ParamType::Struct(types)
-            }
-        }
-    }
-}
-
-fn param_type_calls(field_entries: &[Component]) -> Vec<TokenStream> {
-    field_entries
-        .iter()
-        .map(|Component { field_type, .. }| {
-            let type_name = &field_type.type_name;
-            let parameters = field_type
-                .generic_params
-                .iter()
-                .cloned()
-                .map(TokenStream::from)
-                .collect::<Vec<_>>();
-            if parameters.is_empty() {
-                quote! { <#type_name>::param_type() }
-            } else {
-                quote! { #type_name::<#(#parameters,)*>::param_type() }
-            }
-        })
-        .collect()
-}
-
-fn extract_components(
-    type_decl: &TypeDeclaration,
-    types: &HashMap<usize, TypeDeclaration>,
-    snake_case: bool,
-) -> anyhow::Result<Vec<Component>> {
-    let components = match &type_decl.components {
-        Some(components) if !components.is_empty() => Ok(components),
-        _ => Err(anyhow!(
-            "Custom type {} must have at least one component!",
-            type_decl.type_field
-        )),
-    }?;
-
-    components
-        .iter()
-        .map(|component| Component::new(component, types, snake_case))
-        .collect()
-}
-
-fn extract_generic_parameters(field_types: &[Component]) -> Result<Vec<TokenStream>, LexError> {
-    field_types
-        .iter()
-        .map(|Component { field_type, .. }| field_type.get_used_generic_type_names())
-        .flatten()
-        .unique()
-        .map(|arg| arg.parse())
-        .collect()
-}
-
-// A custom type name should be passed to this function as `{struct,enum} $name`,
-pub(crate) fn extract_custom_type_name_from_abi_property(
-    prop: &TypeDeclaration,
-) -> Result<Ident, Error> {
-    lazy_static! {
-        static ref RE: Regex = Regex::new(r"(?:struct|enum)\s*(.*)").unwrap();
-    }
-
-    RE.captures(&prop.type_field)
-        .map(|captures| ident(&captures[1]))
-        .ok_or_else(|| {
-            Error::InvalidData(
-                "The declared type was not in the format `(enum|struct) name`".to_string(),
-            )
-        })
-}
+pub use common::extract_custom_type_name_from_abi_property;
+pub use enum_gen::expand_custom_enum;
+pub use struct_gen::expand_custom_struct;
 
 // Doing string -> TokenStream -> string isn't pretty but gives us the opportunity to
 // have a better understanding of the generated code so we consider it ok.
